==> 2. data_preprocessing.ipynb <==
### Import libraries

import pandas as pd

### Read dataset

df = pd.read_csv('data/dataset_v1.csv')
df.head(5)

### Data preprocessing

Let's begin by one-hot encoding 'category_id' since its cardinality is 3, which allows us to do this.

cardinality = df.category_id.value_counts(dropna=False).shape[0]
print(f'category_id cardinality: {cardinality}')

df = pd.get_dummies(df, columns=['category_id'], prefix='category')

I believe this was the only variable requiring preprocessing. Let's proceed with the feature engineering phase.

### Export dataframe

df.to_csv('data/dataset_v2.csv', index=False)
